Database now is only dropped if it exists (rather than erroring)

+++ config/descent.sql
@@ -19,7 +19,7 @@ SET time_zone = "+00:00";
 --
 -- Database: `descent`
 --
-DROP DATABASE `descent`;
+IF EXISTS DROP DATABASE `descent`;
 CREATE DATABASE `descent` DEFAULT CHARACTER SET latin1 COLLATE latin1_swedish_ci;
 USE `descent`;
